# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamHub - Community TV Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  admin: "admin",
  operator: "operator",
  user: "user",
} as const;

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  role?: RegisterRequestRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  operator: "operator",
  user: "user",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface Channel {
  id: number;
  name: string;
  description?: string;
  streamKey: string;
  rtmpUrl: string;
  hlsUrl: string;
  webrtcUrl: string;
  isActive: boolean;
  createdById: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateChannelRequest {
  name: string;
  description?: string;
}

export interface UpdateChannelRequest {
  name?: string;
  description?: string;
  isActive?: boolean;
}

export interface ActiveStream {
  id: string;
  name: string;
  channelName?: string;
  clientId?: string;
  vhost?: string;
  app?: string;
  stream?: string;
  bitrate?: number;
  viewers?: number;
  uptime?: number;
  isOnline: boolean;
  videoCodec?: string;
  audioCodec?: string;
  width?: number;
  height?: number;
}

export type StreamClientsClientsItem = {
  id?: string;
  ip?: string;
  type?: string;
  url?: string;
  alive?: number;
};

export interface StreamClients {
  totalClients: number;
  players: number;
  publishers: number;
  clients?: StreamClientsClientsItem[];
}

export interface BandwidthStats {
  ingressBps: number;
  egressBps: number;
  ingressBytes: number;
  egressBytes: number;
}

export type DashboardStatsRecentViewerCountsItem = {
  timestamp?: string;
  count?: number;
};

export interface DashboardStats {
  totalChannels: number;
  activeStreams: number;
  totalViewers: number;
  totalBandwidth: number;
  totalUsers: number;
  totalRevenue: number;
  recentViewerCounts?: DashboardStatsRecentViewerCountsItem[];
}

export type SubscriptionPlanInterval =
  (typeof SubscriptionPlanInterval)[keyof typeof SubscriptionPlanInterval];

export const SubscriptionPlanInterval = {
  monthly: "monthly",
  yearly: "yearly",
} as const;

export interface SubscriptionPlan {
  id: number;
  name: string;
  description?: string;
  price: number;
  interval: SubscriptionPlanInterval;
  maxChannels: number;
  maxBitrate?: number;
  features?: string[];
  isActive: boolean;
  createdAt: string;
}

export type CreatePlanRequestInterval =
  (typeof CreatePlanRequestInterval)[keyof typeof CreatePlanRequestInterval];

export const CreatePlanRequestInterval = {
  monthly: "monthly",
  yearly: "yearly",
} as const;

export interface CreatePlanRequest {
  name: string;
  description?: string;
  price: number;
  interval: CreatePlanRequestInterval;
  maxChannels: number;
  maxBitrate?: number;
  features?: string[];
}

export interface UpdatePlanRequest {
  name?: string;
  description?: string;
  price?: number;
  maxChannels?: number;
  maxBitrate?: number;
  features?: string[];
  isActive?: boolean;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
  cancelled: "cancelled",
} as const;

export interface Invoice {
  id: number;
  userId: number;
  userName?: string;
  subscriptionId?: number;
  amount: number;
  status: InvoiceStatus;
  dueDate: string;
  paidAt?: string | null;
  description?: string;
  createdAt: string;
}

export interface CreateInvoiceRequest {
  userId: number;
  subscriptionId?: number;
  amount: number;
  dueDate: string;
  description?: string;
}

export type UpdateInvoiceRequestStatus =
  (typeof UpdateInvoiceRequestStatus)[keyof typeof UpdateInvoiceRequestStatus];

export const UpdateInvoiceRequestStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
  cancelled: "cancelled",
} as const;

export interface UpdateInvoiceRequest {
  status: UpdateInvoiceRequestStatus;
  paidAt?: string | null;
}

export type SubscriptionStatus =
  (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus];

export const SubscriptionStatus = {
  active: "active",
  cancelled: "cancelled",
  expired: "expired",
} as const;

export interface Subscription {
  id: number;
  userId: number;
  userName?: string;
  planId: number;
  planName?: string;
  channelId?: number;
  channelName?: string;
  status: SubscriptionStatus;
  startDate: string;
  endDate: string;
  createdAt: string;
}

export interface CreateSubscriptionRequest {
  userId: number;
  planId: number;
  channelId?: number;
}

export type ServerStatsCpu = {
  usage?: number;
  cores?: number;
  model?: string;
};

export type ServerStatsMemory = {
  total?: number;
  used?: number;
  free?: number;
  usagePercent?: number;
};

export type ServerStatsNetwork = {
  rxBytes?: number;
  txBytes?: number;
  rxPerSec?: number;
  txPerSec?: number;
};

export interface ServerStats {
  cpu: ServerStatsCpu;
  memory: ServerStatsMemory;
  network: ServerStatsNetwork;
  activeConnections: number;
  uptime: number;
  srsConnected: boolean;
}

export type ListInvoicesParams = {
  userId?: number;
};