# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamHub - Community TV Streaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /billing/plans:
    get:
      operationId: listPlans
      tags: [billing]
      summary: List subscription plans
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Plans list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionPlan"
    post:
      operationId: createPlan
      tags: [billing]
      summary: Create a subscription plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPlan"
  /billing/plans/{id}:
    put:
      operationId: updatePlan
      tags: [billing]
      summary: Update a subscription plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRequest"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPlan"
    delete:
      operationId: deletePlan
      tags: [billing]
      summary: Delete a subscription plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /billing/invoices:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [billing]
      summary: Create an invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/invoices/{id}:
    put:
      operationId: updateInvoice
      tags: [billing]
      summary: Update invoice status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /billing/subscriptions:
    get:
      operationId: listSubscriptions
      tags: [billing]
      summary: List subscriptions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [billing]
      summary: Create a subscription (channel rental)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        role:
          type: string
          enum: [admin, operator, user]
          default: user
      required:
        - email
        - password
        - name
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, operator, user]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
...
    SubscriptionPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        interval:
          type: string
          enum: [monthly, yearly]
        maxChannels:
          type: integer
        maxBitrate:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - interval
        - maxChannels
        - isActive
        - createdAt
    CreatePlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        interval:
          type: string
          enum: [monthly, yearly]
        maxChannels:
          type: integer
        maxBitrate:
          type: integer
        features:
          type: array
          items:
            type: string
      required:
        - name
        - price
        - interval
        - maxChannels
    UpdatePlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        maxChannels:
          type: integer
        maxBitrate:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    Invoice:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        subscriptionId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, overdue, cancelled]
        dueDate:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amount
        - status
        - dueDate
        - createdAt
    CreateInvoiceRequest:
      type: object
      properties:
        userId:
          type: integer
        subscriptionId:
          type: integer
        amount:
          type: number
        dueDate:
          type: string
          format: date-time
        description:
          type: string
      required:
        - userId
        - amount
        - dueDate
    UpdateInvoiceRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, overdue, cancelled]
        paidAt:
          type: string
          format: date-time
          nullable: true
      required:
        - status
    Subscription:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        planId:
          type: integer
        planName:
          type: string
        channelId:
          type: integer
        channelName:
          type: string
        status:
          type: string
          enum: [active, cancelled, expired]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - planId
        - status
        - startDate
        - endDate
        - createdAt
    CreateSubscriptionRequest:
      type: object
      properties:
        userId:
          type: integer
        planId:
          type: integer
        channelId:
          type: integer
      required:
        - userId
        - planId
...